refactor: 상품 및 프로모션 데이터 등록 dto 패키지 변경
