+++ src/main/java/store/controller/ConvenienceController.java
@@ -3,8 +3,8 @@ package store.controller;
 import java.util.List;
 import store.dto.AdditionalQuantityRequest;
 import store.dto.OrderRequest;
-import store.dto.ProductInput;
-import store.dto.PromotionInput;
+import store.dto.data.ProductInput;
+import store.dto.data.PromotionInput;
 import store.dto.RemoveNonPromotionRequest;
 import store.exception.ExceptionStatus;
 import store.exception.InvalidPromotionDateException;
