+++ src/main/java/store/service/ConvenienceService.java
@@ -4,9 +4,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import store.dto.OrderRequest;
-import store.dto.ProductInput;
+import store.dto.data.ProductInput;
 import store.dto.ProductShowResponse;
-import store.dto.PromotionInput;
+import store.dto.data.PromotionInput;
 import store.exception.NoProductException;
 import store.model.Order;
 import store.model.Product;
