+++ src/test/java/store/repository/ProductRepositoryTest.java
@@ -5,8 +5,8 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import store.dto.ProductInput;
-import store.dto.PromotionInput;
+import store.dto.data.ProductInput;
+import store.dto.data.PromotionInput;
 import store.exception.DuplicatePromotionException;
 import store.model.Product;
 import store.model.Promotion;
