+++ src/test/java/store/service/ConvenienceServiceTest.java
@@ -9,8 +9,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import store.dto.OrderRequest;
-import store.dto.ProductInput;
-import store.dto.PromotionInput;
+import store.dto.data.ProductInput;
+import store.dto.data.PromotionInput;
 import store.model.Order;
 import store.model.Product;
 import store.model.Promotion;
